resolve, add global cache
