+++ src/btree/multi.ts
@@ -5,20 +5,22 @@ import { PageFile } from "./pagefile";
 const PAGE_SIZE_BYTES = 4096;
 export const maxUint64 = 2n ** 64n - 1n;
 
+const globalMetaPageCache = new Map<
+  bigint,
+  Promise<{ data: ArrayBuffer; totalLength: number }[]>
+>();
+
 export class LinkedMetaPage {
   private resolver: RangeResolver;
   private offset: bigint;
-  private metaPageData: ArrayBuffer | null;
-  private metaPagePromise: Promise<ArrayBuffer> | null = null;
-
   constructor(resolver: RangeResolver, offset: bigint) {
     this.resolver = resolver;
     this.offset = offset;
-    this.metaPageData = null;
   }
 
   async root(): Promise<MemoryPointer> {
-    const pageData = await this.getMetaPage();
+    const res = await this.getMetaPage();
+    const { data: pageData } = res[0];
 
     // we seek by 12 bytes since offset is 8 bytes, length is 4 bytes
     const data = pageData.slice(0, 12);
@@ -40,7 +42,8 @@ export class LinkedMetaPage {
    * 		(3) slices from [24, (24 + dataLength)] which contain metadata
    */
   async metadata(): Promise<ArrayBuffer> {
-    const pageData = await this.getMetaPage();
+    const res = await this.getMetaPage();
+    const { data: pageData } = res[0];
 
     const lengthView = new DataView(pageData, 24);
 
@@ -50,42 +53,29 @@ export class LinkedMetaPage {
     return pageData.slice(28, 28 + metadataLength);
   }
 
-  /**
-   * `getMetaPage()` seeks the index-file with the absolute bounds for a given page file.
-   * It caches the data in a pagefile. Note: all other methods that call this should be slicing with relative bounds.
-   */
-  private async getMetaPage(): Promise<ArrayBuffer> {
-    if (this.metaPageData) {
-      return this.metaPageData;
-    }
-
-    if (!this.metaPagePromise) {
-      this.metaPagePromise = this.resolver([
+  private async getMetaPage(): Promise<
+    { data: ArrayBuffer; totalLength: number }[]
+  > {
+    if (!globalMetaPageCache.has(this.offset)) {
+      const p = this.resolver([
         {
           start: Number(this.offset),
           end: Number(this.offset) + PAGE_SIZE_BYTES - 1,
         },
-      ])
-        .then((r) => {
-          const { data } = r[0];
-          this.metaPageData = data;
-          this.metaPagePromise = null;
-          return data;
-        })
-        .catch((error) => {
-          this.metaPagePromise = null;
-          throw error;
-        });
+      ]);
+
+      globalMetaPageCache.set(this.offset, p);
     }
 
-    return this.metaPagePromise;
+    return globalMetaPageCache.get(this.offset)!;
   }
 
   /**
    * `next()` - returns a new LinkedMetaPage
    */
   async next(): Promise<LinkedMetaPage | null> {
-    const pageData = await this.getMetaPage();
+    const res = await this.getMetaPage();
+    const { data: pageData } = res[0];
 
     const view = new DataView(pageData, 12, 8);
     const nextOffset = view.getBigUint64(0, true);
